% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestModel.r
\name{getBestModel}
\alias{getBestModel}
\title{Get Best possible model for your binary class dataset}
\usage{
getBestModel(x, y, iter = 100, verboseIter = FALSE)
}
\arguments{
\item{x}{Matrix or dataframe with numeric values (features of train data)}

\item{y}{Vector (class of given train data)}

\item{iter}{Numeric value (These many number of time each model will be evaluated for AUC)}

\item{verboseIter}{Boolean value (print progress)}
}
\value{
dataframe with fields for Average AUC, Model Name, Platform (Ex. r)
}
\description{
Gives the best possible binary classification model suit for given dataset.
Takes binary class train dataset and it's labels, perform stratified sampling,
performs train and test, gives average AUC for 'iter' iterations
}

