% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPredict.r
\name{getLastModel}
\alias{getLastModel}
\title{Get last trained model using getTrainModel()}
\usage{
getLastModel()
}
\value{
Vector that contains environment and algorithm name
}
\description{
This funtion returns the info of last recent trained model
i.e. environment and algorithm name
}

