% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPredict.r
\name{getPredictProb}
\alias{getPredictProb}
\title{Get prediction in probabilty}
\usage{
getPredictProb(model, testData)
}
\arguments{
\item{model}{trained model with getTrainModel function}

\item{testData}{Test dataset having same number of columns as test dataset used to trained model}
}
\value{
Dataframe with probabilities for both classes
}
\description{
This funtion predict class of given test data in terms of probability
}

