% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPredict.r
\name{getTrainModel}
\alias{getTrainModel}
\title{Get trained model for given binary class method name}
\usage{
getTrainModel(x, y, methodName, verboseIter = FALSE, tuneLength = 10)
}
\arguments{
\item{x}{Matrix or dataframe with numeric values (features of train data)}

\item{y}{Vector (class of given train data)}

\item{methodName}{Dataframe row returned from getBestModel(Ex. method_list[1,], here method_list is a dataframe returned from getBestModel and giving 1st method as input to train model)}

\item{verboseIter}{Boolean value (print progress)}
}
\value{
Vector that contains environment and algorithm name
}
\description{
This funtion performs the hyper-tuning for given model.
Trains model on full train dataset, and returns trained model
}

